+++ src/components/details/Details.tsx
@@ -1,4 +1,4 @@
-import { CurrentNews } from "./current-news/CurrentNews";
+import { CurrentNews } from "./components/current-news/CurrentNews";
 
 export const Details: React.FC = () => {
   return <CurrentNews />;
