+++ src/components/details/components/news-details/NewsDetails.tsx
